Merge to update 0.2

+++ CSClock/Program.cs
@@ -63,6 +63,8 @@ namespace CSClock
         public static ResourceManager rm_SetOvertime = null;
         public static ResourceManager rm_AddSubtractTime = null;
 
+        private static bool portable = true;
+
         /// <summary>
         /// The main entry point for the application.
         /// </summary>
@@ -75,7 +77,14 @@ namespace CSClock
 #if DEBUG
             debug = true;
 #endif
+
+            if (args != null && args.Length > 0 && args.Contains("-np"))
+            {
+                portable = false;
+            }
+
             string logDir = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "CSClock");
+
             if (!Directory.Exists(logDir))
             {
                 Directory.CreateDirectory(logDir);
@@ -90,7 +99,6 @@ namespace CSClock
                 }
             }
 
-            
             logger = new Logger("CSClock", Path.Combine(logDir, "log.txt"),
                 Logger.LogTimeDateOptions.YearMonthDayHourMinuteSecond, true);
 
@@ -99,6 +107,10 @@ namespace CSClock
             {
                 if (createdNew || (args != null && args.Contains("-ignorerunning")))
                 {
+                    if (!portable)
+                    {
+                        Update();
+                    }
                     logger.Log(className, "createdNew=true", Logger.LogType.Info, true);
                     StartApplication(args);
                 }
@@ -135,6 +147,28 @@ namespace CSClock
             }
         }
 
+        static void Update()
+        {
+            try
+            {
+                string updaterPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "CSClock", "Updater.exe");
+                ProcessStartInfo start =
+                    new ProcessStartInfo();
+                start.FileName = updaterPath;
+                start.Arguments = "-update";
+                start.WindowStyle = ProcessWindowStyle.Hidden;
+                var updateProc = Process.Start(start);
+                updateProc.WaitForExit();
+            }
+            catch (Exception ex)
+            {
+                logger.Log(className, "Error when running auto-updater at app launch: " + ex.ToString(), Logger.LogType.Error);
+                MessageBox.Show("Error when running auto-updater: " + ex.Message + "\r\n\r\nSee '" + Path.Combine(Environment.GetFolderPath(
+                    Environment.SpecialFolder.LocalApplicationData), @"CSClock\log.txt") + "' for more details", "CSClock",
+                    MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+        }
+
         static void StartApplication(string[] args)
         {
             assembly = Assembly.GetExecutingAssembly();
@@ -142,7 +176,7 @@ namespace CSClock
 
             logger.Log(className, "Starting CSClock", Logger.LogType.Info);
 
-            if (Properties.Settings.Default.upgradeRequired && (args == null || !args.Contains("-donotupgradesettings")))
+            if (Properties.Settings.Default.upgradeRequired && ((args == null && args.Length > 0) || !args.Contains("-donotupgradesettings")))
             {
                 logger.Log(className, "Upgrading user settings...", Logger.LogType.Info);
 
@@ -157,13 +191,13 @@ namespace CSClock
             rm_Messages = new ResourceManager(string.Format("CSClock.Languages.{0}.Messages", selectedLanguage), assembly);
             rm_GUI = new ResourceManager(string.Format("CSClock.Languages.{0}.GUI", selectedLanguage), assembly);
 
-            if (args != null && args.Contains("-removal"))
+            if (args != null && args.Length > 0 && args.Contains("-removal"))
             {
                 Removal(true);
                 return;
             }
 
-            if (args != null && args.Contains("-reset"))
+            if (args != null && args.Length > 0 && args.Contains("-reset"))
             {
                 logger.Log(className, "Resetting CSClock", Logger.LogType.Info);
 
@@ -173,13 +207,13 @@ namespace CSClock
 
                 logger.Log(className, "Reset completed, closing CSClock", Logger.LogType.Info);
 
-                if ((args == null || !args.Contains("-deletelog")) || !File.Exists("log.txt"))
+                if (((args == null && args.Length > 0) || !args.Contains("-deletelog")) || !File.Exists("log.txt"))
                 {
                     return;
                 }
             }
 
-            if (args != null && args.Contains("-deletelog") && File.Exists("log.txt"))
+            if (args != null && args.Length > 0 && args.Contains("-deletelog") && File.Exists("log.txt"))
             {
                 File.Delete("log.txt");
                 return;
@@ -263,7 +297,9 @@ namespace CSClock
             }
             notifyIcon1 = new NotifyIcon();
             notifyIcon1.Icon = Properties.Resources.Logo;
+            notifyIcon1.DoubleClick += ShowGUI;
             contextMenu1 = new ContextMenu();
+            contextMenu1.MenuItems.Add(rm_GUI.GetString("showGUI"), new EventHandler(ShowGUI));
             if (Properties.Settings.Default.timerEnabled)
             {
                 contextMenu1.MenuItems.Add(rm_GUI.GetString("pauseResume"), new EventHandler(PauseResumeTimer));
@@ -290,6 +326,7 @@ namespace CSClock
         {
             if (CSClockForm != null)
             {
+                CSClockForm.Show();
                 CSClockForm.BringToFront();
             }
             else
